requirements.txt
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-gis==0.17
gunicorn==19.9.0
python-decouple==3.4
whitenoise==5.3.0